backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-docx
fpdf
python-multipart
Jinja2
